New empty privacy manifest
The previous privacy manifest has been rejected by Apple. I tested it. It fails the automatic checks when submitting a build for review. This was their response:

> ITMS-91056: Invalid privacy manifest - The PrivacyInfo.xcprivacy file from the following path is invalid:
> “StatKit_StatKit.bundle/PrivacyInfo.xcprivacy”. Keys and values in your app’s privacy manifests must be valid. For more
> details about privacy manifest files, visit:
> https://developer.apple.com/documentation/bundleresources/privacy_manifest_files.

This new privacy manifest gives the same information as far as I can see it: No tracking, nothing. It’s "empty".

This new privacy manifest has been accepted by Apple on Sunday, May 18th 2025. It passes both the automatic and the manual review.

Signed-off-by: Marian Stiehler <mail@stiehler.dev>